requirements.txt
torch
transformers==4.40.2
PyTDC
scikit-learn
scipy
requests
rdkit
xgboost
minimol
chemprop
setuptools<71